backend/requirements.txt
requests
camelot-py
opencv-python
ghostscript
fastapi
uvicorn
python-multipart
